// src/shared/models/http-response.models.ts
export enum HTTP_STATUS {
  //
  OK = 200,
  CREATED = 201,
  NO_CONTENT = 204,
  //
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  PAYMENT_REQUIRED = 402,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  REQUEST_TIMEOUT = 408,
  CONFLICT = 409,
  //
  INTERNAL_SERVER_ERROR = 500,
  SERVICE_UNAVAILABLE = 503
}